Add commands work and example work

+++ extract_params.py
@@ -4,7 +4,7 @@
 The parameter file can be used to add a new equal simulation, but more usefully
 modified to run a slightly different simulation.
 """
-# Copyright (C) 2017 Håkon Austlid Taskén <hakon.tasken@gmail.com>
+# Copyright (C) 2017, 2018 Håkon Austlid Taskén <hakon.tasken@gmail.com>
 # Licenced under the MIT License.
 
 import helpers
@@ -12,11 +12,17 @@ import sqlite3
 import argparse
 import os.path
 
+no_extract_columns = {'id', 'status', 'comment', 'time_submitted', 'time_started', \
+                      'used_walltime', 'job_id', 'cpu_info', 'git_hash', \
+                      'commit_message', 'git_diff_stat', 'git_diff', 'sha1_executables'}
+
 def get_arguments(argv):
-    parser = argparse.ArgumentParser(description='Extract parameter file from sim_runs.db.')
-    parser.add_argument('-id', type=int, required=True, help="<Required> ID of the simulation which parameter one wish to extract.")
-    parser.add_argument('-filename', '-f', type=str, default=None, help="Name of paramter file generated.")
-    parser.add_argument('--all', action='store_true', help="Extract all parameters. Default if to only extract non empty parameters to parameter file.")
+    parser = argparse.ArgumentParser(description='Extract parameter file from sim.db.')
+    parser.add_argument('--id', '-i', type=int, required=True, help="<Required> ID of the simulation which parameter one wish to extract.")
+    parser.add_argument('--filename', '-f', type=str, default=None, help="Name of paramter file generated.")
+    parser.add_argument('--default_file', '-d', action='store_true', help="Write parameters to the first of the 'Parameter filenames' in settings.txt. Ask for confirmation if file exists already.")
+    parser.add_argument('--also_empty', action='store_true', help="Also extract empty paramters. Default is to not extract empty parameters and default columns that are not input parameters.")
+    parser.add_argument('--all', action='store_true', help="Extract all parameters. Default is to not extract empty parameters and default columns that are not input parameters.")
     return parser.parse_args(argv)
 
 def get_param_type_as_string(col_type, value):
@@ -38,14 +44,28 @@ def get_param_type_as_string(col_type, value):
 def extract_params(argv=None):
     args = get_arguments(argv)
 
-    if args.filename == None:
-        filename = 'sim_params.txt'
-    else:
+    is_printing_parameters = True
+    if args.default_file:
+        param_files = helpers.Settings().read('parameter_files')
+        if len(param_files) == 0:
+            print("ERROR: No '--filename' provided and no 'Parameters files' in settings.txt.")
+            exit()
+        else:
+            filename = param_files[0]
+        if os.path.exists(filename):
+            answer = helpers.user_input("Would you like to overwrite '{}'? (y/n)" \
+                    .format(filename))
+            if answer != 'y' and answer != 'Y' and answer != 'yes' and answer != 'Yes':
+                exit()
+        print("Extracts parameters to '{}'.".format(filename))
+        is_printing_parameters = False
+    elif args.filename != None:
         filename = args.filename
-    params_file = open(filename, 'w')
+        is_printing_parameters = False
+    if not is_printing_parameters:
+        params_file = open(filename, 'w')
 
-    sim_db_dir = helpers.get_closest_sim_db_path()
-    db = sqlite3.connect(sim_db_dir + 'sim.db')
+    db = helpers.connect_sim_db()
     db_cursor = db.cursor()
 
     db_cursor.execute("SELECT * FROM runs WHERE id={}".format(args.id))
@@ -55,22 +75,24 @@ def extract_params(argv=None):
 
     for col_name, col_type, value in zip(column_names, column_types, 
                                          extracted_row[0]):
-        if col_name != 'id' and col_name != 'status' \
-                and col_name != 'used_walltime' and col_name != 'job_id':
+        if col_name in no_extract_columns:
             skip = True
         else:
             skip = False
 
-        if (args.all or value != None) and skip:
+        if ((args.also_empty or value != None) and not skip) or args.all:
             line = col_name
             param_type = get_param_type_as_string(col_type, value)
             line += " ({}): ".format(param_type)
             if param_type[-3:] == 'ray':
                 value = '[' + value.split('[')[1]
-            line += str(value) + '\n'
-            params_file.write(line)
-    
-    params_file.close()        
+            line += str(value).replace(':', ';') + '\n'
+            if is_printing_parameters:
+                print(line)
+            else:
+                params_file.write(line)
+    if not is_printing_parameters:
+        params_file.close()        
 
     db.commit()
     db_cursor.close()
